requirements.txt
alabaster==0.7.6
appnope==0.1.0
Babel==2.1.1
backports.ssl-match-hostname==3.4.0.2
BeautifulSoup==3.2.1
bitcoin==1.1.42
certifi==2015.9.6.2
click==5.1
cssselect==0.9.1
cycler==0.9.0
decorator==4.0.9
Django==1.8.5
docopt==0.6.2
docutils==0.12
eventlet==0.18.4
first==2.0.1
Flask==0.10.1
funcsigs==0.4
functools32==3.2.3.post2
gdbn==0.1
gnumpy==0.2
gnureadline==6.3.3
greenlet==0.4.9
html5lib==0.9999999
ipykernel==4.1.0
ipyparallel==4.0.2
ipython==4.1.2
ipython-genutils==0.1.0
isodate==0.5.4
itsdangerous==0.24
Jinja2==2.8
joblib==0.9.2
jsonschema==2.5.1
jupyter-client==4.1.1
jupyter-core==4.0.6
Lasagne==0.1
lxml==3.4.4
MarkupSafe==0.23
matplotlib==1.5.0
mercurial==3.7.1
mistune==0.7.1
mock==1.3.0
nbconvert==4.0.0
nbformat==4.0.1
networkx==1.10
nltk==3.1
nolearn==0.5
nose==1.3.7
notebook==4.0.6
numpy==1.10.4
numpydoc==0.5
pandas==0.17.0
path.py==8.1.2
pbr==1.8.1
pdfminer==20140328
pexpect==4.0.1
pickleshare==0.6
Pillow==3.0.0
pip-tools==1.1.6
prettyprint==0.1.5
protobuf==2.6.1
ptyprocess==0.5.1
Pygments==2.0.2
pygobject==3.18.2
pylab==0.1.3
pyparsing==2.0.3
pyquery==1.2.9
pyreadline==2.1
python-dateutil==2.4.2
python-engineio==0.8.8
python-nmap==0.5.0.post1
python-socketio==1.0
pytz==2015.6
pyzmq==14.7.0
q==2.6
qtconsole==4.1.0
rdflib==4.2.1
requests==2.8.1
scikit-image==0.11.3
scikit-learn==0.16.1
scipy==0.16.0
seaborn==0.6.0
simplegeneric==0.8.1
six==1.10.0
snakeviz==0.4.0
snowballstemmer==1.2.0
SPARQLWrapper==1.6.4
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
sympy==0.7.6.1
terminado==0.5
testpath==0.2
Theano==0.7.0
tornado==4.2.1
traitlets==4.1.0
virtualenv==13.1.2
Werkzeug==0.11.4
wheel==0.29.0
